Organized components into respective folders
The folder structure was not looking good and i thought of organizing the files and folders which would make easier and readable.

Before it was a component folder and inside it, there was a "Body" Folder and the body folder has all the necessary components of the application.

It was getting very difficult to read and I made Folders like footer, header, MenuList, HomePage, AdminControls, UserControls. And I moved all the respective files to the corresponding folders